Moved Fts into their own indiv files
